requirements.txt
Flask==2.0.3
line-bot-sdk==2.2.1
matplotlib==3.1.1
requests==2.23.0
requests-oauthlib==1.3.0
urllib3==1.25.3